# codegen/src/model/EModelEvent.js
/*
 * Gantt Chart Schedule API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 *
 * Swagger Codegen version: 3.0.68
 *
 * Do not edit the class manually.
 *
 */
import ApiClient from '../ApiClient.js';

/**
 * Enum class EModelEvent.
 * @enum {String}
 * @readonly
 */
const EModelEvent = {
  /**
   * value: "ADD"
   * @const
   */
  ADD: "ADD",

  /**
   * value: "CHANGE"
   * @const
   */
  CHANGE: "CHANGE",

  /**
   * value: "REMOVE"
   * @const
   */
  REMOVE: "REMOVE",

  /**
   * Returns a <code>EModelEvent</code> enum value from a JavaScript object name.
   * @param {Object} data The plain JavaScript object containing the name of the enum value.
  * @return {module:model/EModelEvent} The enum <code>EModelEvent</code> value.
   */
  constructFromObject: function(object) {
    return object;
  }
};

export default {EModelEvent};