# codegen/src/model/GanttChartModelEvent.js
/*
 * Gantt Chart Schedule API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 *
 * Swagger Codegen version: 3.0.68
 *
 * Do not edit the class manually.
 *
 */
import ApiClient from '../ApiClient.js';
import EModelEvent from './EModelEvent.js';
import ScheduledTask from './ScheduledTask.js';

/**
 * The GanttChartModelEvent model module.
 * @module model/GanttChartModelEvent
 * @version 1.0.0
 */
export default class GanttChartModelEvent {
  /**
   * Constructs a new <code>GanttChartModelEvent</code>.
   * @alias module:model/GanttChartModelEvent
   * @class
   * @param modelEvent {module:model/EModelEvent} 
   */
  constructor(modelEvent) {
    this.modelEvent = modelEvent;
  }

  /**
   * Constructs a <code>GanttChartModelEvent</code> from a plain JavaScript object, optionally creating a new instance.
   * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
   * @param {Object} data The plain JavaScript object bearing properties of interest.
   * @param {module:model/GanttChartModelEvent} obj Optional instance to populate.
   * @return {module:model/GanttChartModelEvent} The populated <code>GanttChartModelEvent</code> instance.
   */
  static constructFromObject(data, obj) {
    if (data) {
      obj = obj || new GanttChartModelEvent();
      if (data.hasOwnProperty('modelEvent'))
        obj.modelEvent = EModelEvent.constructFromObject(data['modelEvent']);
      if (data.hasOwnProperty('oldTask'))
        obj.oldTask = ScheduledTask.constructFromObject(data['oldTask']);
      if (data.hasOwnProperty('newTask'))
        obj.newTask = ScheduledTask.constructFromObject(data['newTask']);
    }
    return obj;
  }
}

/**
 * @member {module:model/EModelEvent} modelEvent
 */
GanttChartModelEvent.prototype.modelEvent = undefined;

/**
 * @member {module:model/ScheduledTask} oldTask
 */
GanttChartModelEvent.prototype.oldTask = undefined;

/**
 * @member {module:model/ScheduledTask} newTask
 */
GanttChartModelEvent.prototype.newTask = undefined;