# codegen/src/model/ScheduledTask.js
/*
 * Gantt Chart Schedule API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 *
 * Swagger Codegen version: 3.0.68
 *
 * Do not edit the class manually.
 *
 */
import ApiClient from '../ApiClient.js';
import SerializableColor from './SerializableColor.js';

/**
 * The ScheduledTask model module.
 * @module model/ScheduledTask
 * @version 1.0.0
 */
export default class ScheduledTask {
  /**
   * Constructs a new <code>ScheduledTask</code>.
   * @alias module:model/ScheduledTask
   * @class
   * @param id {Number} 
   */
  constructor(id) {
    this.id = id;
  }

  /**
   * Constructs a <code>ScheduledTask</code> from a plain JavaScript object, optionally creating a new instance.
   * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
   * @param {Object} data The plain JavaScript object bearing properties of interest.
   * @param {module:model/ScheduledTask} obj Optional instance to populate.
   * @return {module:model/ScheduledTask} The populated <code>ScheduledTask</code> instance.
   */
  static constructFromObject(data, obj) {
    if (data) {
      obj = obj || new ScheduledTask();
      if (data.hasOwnProperty('id'))
        obj.id = ApiClient.convertToType(data['id'], 'Number');
      if (data.hasOwnProperty('groupName'))
        obj.groupName = ApiClient.convertToType(data['groupName'], 'String');
      if (data.hasOwnProperty('name'))
        obj.name = ApiClient.convertToType(data['name'], 'String');
      if (data.hasOwnProperty('startTime'))
        obj.startTime = ApiClient.convertToType(data['startTime'], 'Date');
      if (data.hasOwnProperty('endTime'))
        obj.endTime = ApiClient.convertToType(data['endTime'], 'Date');
      if (data.hasOwnProperty('colour'))
        obj.colour = SerializableColor.constructFromObject(data['colour']);
    }
    return obj;
  }
}

/**
 * @member {Number} id
 */
ScheduledTask.prototype.id = undefined;

/**
 * @member {String} groupName
 */
ScheduledTask.prototype.groupName = undefined;

/**
 * @member {String} name
 */
ScheduledTask.prototype.name = undefined;

/**
 * @member {Date} startTime
 */
ScheduledTask.prototype.startTime = undefined;

/**
 * @member {Date} endTime
 */
ScheduledTask.prototype.endTime = undefined;

/**
 * @member {module:model/SerializableColor} colour
 */
ScheduledTask.prototype.colour = undefined;