# codegen/src/model/SerializableColor.js
/*
 * Gantt Chart Schedule API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 *
 * Swagger Codegen version: 3.0.68
 *
 * Do not edit the class manually.
 *
 */
import ApiClient from '../ApiClient.js';

/**
 * The SerializableColor model module.
 * @module model/SerializableColor
 * @version 1.0.0
 */
export default class SerializableColor {
  /**
   * Constructs a new <code>SerializableColor</code>.
   * @alias module:model/SerializableColor
   * @class
   * @param red {Number} 
   * @param green {Number} 
   * @param blue {Number} 
   * @param alpha {Number} 
   */
  constructor(red, green, blue, alpha) {
    this.red = red;
    this.green = green;
    this.blue = blue;
    this.alpha = alpha;
  }

  /**
   * Constructs a <code>SerializableColor</code> from a plain JavaScript object, optionally creating a new instance.
   * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
   * @param {Object} data The plain JavaScript object bearing properties of interest.
   * @param {module:model/SerializableColor} obj Optional instance to populate.
   * @return {module:model/SerializableColor} The populated <code>SerializableColor</code> instance.
   */
  static constructFromObject(data, obj) {
    if (data) {
      obj = obj || new SerializableColor();
      if (data.hasOwnProperty('red'))
        obj.red = ApiClient.convertToType(data['red'], 'Number');
      if (data.hasOwnProperty('green'))
        obj.green = ApiClient.convertToType(data['green'], 'Number');
      if (data.hasOwnProperty('blue'))
        obj.blue = ApiClient.convertToType(data['blue'], 'Number');
      if (data.hasOwnProperty('alpha'))
        obj.alpha = ApiClient.convertToType(data['alpha'], 'Number');
    }
    return obj;
  }
}

/**
 * @member {Number} red
 */
SerializableColor.prototype.red = undefined;

/**
 * @member {Number} green
 */
SerializableColor.prototype.green = undefined;

/**
 * @member {Number} blue
 */
SerializableColor.prototype.blue = undefined;

/**
 * @member {Number} alpha
 */
SerializableColor.prototype.alpha = undefined;